barbershop/products/serializers.py
from rest_framework import serializers
from .models import Catalog, Product


class CatalogSerializer(serializers.ModelSerializer):
    class Meta:
        model = Catalog
        fields = ['name', 'slug']


class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = Product
        fields = ['id', 'name', 'slug', 'price', 'stock', 'available', 'created', 'updated']
