barbershop/requirements.txt
Django==2.2.0
django-extensions==3.0.3
django-filter==2.3.0
django-jet2==1.0.12
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.0
django-novaposhta


